add ftw.simplelayout support.
This automatically adds support for ftw.simplelayout pages when the
package is installed.
When a page is published, all contained blocks without a workflow
are published as well.

The back-references checks, which are done for the old archetypes
simplelayout implementation, are not yet done, because this does not
work well with dexterity.