do not publish delete blocks for ftw.simplelayout blocks.
ftw.simplelayout blocks are removed on the receiver side when the page
is published and the block is no longer in the page config.
Therefore the block should not be deleted with a delete job.

This makes it possible to remove blocks without having them removed too
early from the receiver side.
The idea here is that the block is part of the page's content and not
seen as separate, standalone object.