No longer move simplelayout blocks instantly.
Simplelayout blocks are considered part of the content and have no
workflow for theirself.
Therefore we usually want to publish changes in the blocks along with
the page.

But when cutting / pasting blocks from one page to another, these
changes were instantly published, which was too early.

Since the simplelayout publisher adapter makes sure that blocks which no
longer exist on the backend are removed from the frontend when
publishing the parent page, we can stop moving simplelayout blocks
instantly.