Fix extraction token access error on first publishing.
The job of the extraction token is to invoke a retry of the job
execution when the base transaction does not contain the changes which
should be published.

This edge case can happen on slow systems when the taskqueue worker is
faster than the ZODB, probably in combination with a multi instance
setup.

The problem is that when publishing the object the first time, the token
is not yet existing on the object when we have on old transaction.
The keyword access caused a KeyError at this point.