Fix extraction bug on move when redis is active.
When redis is active for deferred extraction in a multi instance setup,
the move job may contain invalid paths when the extraction job is
processed in a different process than the job was created.

The problem is that the event infos are stored on the object until the
extraction happened. Since the event infos contain acquisition wrapped
objects and we deferreing results in that the extraction happens in a
different transaction, it caused persistent pointers to be stored in the
database.
The extraction did rely on having acquisition wrapped objects, which was
sometimes not the case, depending on the database cache state.

The solution is to not store anything and prepare strings early.