Make renaming objects on the Plone Site work again.
The assert statement's condition did not support the case when the `obj` is the site root. This situation can happen when renaming objects on the site root.

The condition's job was to erify that objects actually are stored within the portal. But since the portal object is found by walking up the object acquisition chain, we are already sure that the object is stored within the portal object. Therefore the assertion is dumb, unnecessary and wrong.