collective.easyform safe data download support.
The goal is that downloading the safe data adapter data on the
editorial site will also include the data of the public facing site,
because editors usually cannot log into the public facing site.

The patch combines the data from the sender installation and the receiver
installation to one csv / tsv. We assume that the form config is the
same on both sides

The concept is to first execute the standard implementation, which sets the
response headers and streams the local data.

If the context is considered public, the second step is to get the data from
the remote realms and append it to the response body (with response.write).

This will combine the data of the all sites in a streamed http response.

This implementation also works in local development with two plone sites,
assuming that the receiver side does not have any realms configured.